refactor: article rest controller
